+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Return trace ID in error responses [#13](https://github.com/rokwire/logging-library-go/issues/13)
+- Add OpenShift health check request properties builder for log suppression [#14](https://github.com/rokwire/logging-library-go/issues/14)
 
 ## [1.0.3] - 2021-12-02
 ### Fixed